% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nhance.R
\name{find_nhance}
\alias{find_nhance}
\title{find_nhance}
\usage{
find_nhance(year_begin, year_end)
}
\arguments{
\item{year_begin}{the year begin}

\item{year_end}{the year end}
}
\value{
title
}
\description{
find nhance data and download it.
}
\author{
fanchenzashi
}
